Split common.h into multiple .h files, named after what is actually in them.
